+++ src/api/utils.ts
@@ -12,26 +12,13 @@ import { AnyNumber } from "../types";
 export async function waitForIndexerOnVersion(args: {
   config: AptosConfig;
   minimumLedgerVersion?: AnyNumber;
-  processorTypes: Array<ProcessorType>;
+  processorType: ProcessorType;
 }) {
   if (args.minimumLedgerVersion !== undefined) {
-    // eslint-disable-next-line no-restricted-syntax
-
-    // Collect all of the promises to wait at the same time
-    // TODO(greg): This needs to be refactored to be handled at the query time.  It involves rewriting
-    // a bunch of the logic for wait for indexer
-    const promises: Array<Promise<void>> = [];
-    // eslint-disable-next-line no-restricted-syntax
-    for (const processorType of args.processorTypes) {
-      promises.push(
-        waitForIndexer({
-          aptosConfig: args.config,
-          minimumLedgerVersion: args.minimumLedgerVersion,
-          processorType,
-        }),
-      );
-    }
-
-    await Promise.all(promises);
+    await waitForIndexer({
+      aptosConfig: args.config,
+      minimumLedgerVersion: args.minimumLedgerVersion,
+      processorType: args.processorType,
+    });
   }
 }